relocate xslt related classes into scripting contrib
