+++ solr/core/src/java/org/apache/solr/handler/UpdateRequestHandler.java
@@ -168,7 +168,7 @@ public class UpdateRequestHandler extends ContentStreamHandlerBase implements Pe
 
   @Override
   public String getDescription() {
-    return "Add documents using XML (with XSLT), CSV, JSON, or javabin";
+    return "Add documents using XML, CSV, JSON, or javabin.";
   }
 
   @Override
@@ -182,6 +182,3 @@ public class UpdateRequestHandler extends ContentStreamHandlerBase implements Pe
   public static final String BIN_PATH = "/update/bin";
 
 }
-
-
-
